backend/pyproject.toml
[tool.poetry]
name = "study-it"
version = "0.1.0"
description = "study python"
authors = ["artrmii"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11, <4.0"
fastapi = {extras = ["standard"], version = "^0.116"}
uvicorn = {extras = ["standard"], version = "^0.35"}
django-split-settings = "^1.1.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.25"}
aerich = "^0.9"
httpx = "^0.28"
pydantic-settings = "^2.10"
sentry-sdk = "^2.37"
pydantic-i18n = "^0.4.5"
fastapi-babel = "^1"
fastapi-pagination = "^0.14"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.2"
requests = "^2.28.1"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"

[tool.aerich]
tortoise_orm = "config.tortoise_settings"
location = "app/db/migrations"
src_folder = "./app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "app",
  ".",
]
addopts = [
    "--import-mode=importlib",
]
asyncio_default_fixture_loop_scope = "module"
